refactor: convert dag API to async/await
